requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pydantic
pydantic-settings
firebase
firebase-admin
